Rename lucina python package to minimd
