--- lucina/__init__.py
@@ -1,12 +1,0 @@
-from lucina._about import __author__
-from lucina._about import __author_email__
-from lucina._about import __license__
-from lucina._about import __url__
-from lucina._about import __version__
-from lucina.tokenizer import tokenize_files
-
-
-__all__ = [
-    '__author__', '__author_email__', '__license__', '__url__', '__version__',
-    'tokenize_files',
-]
