+++ minimd/__init__.py
@@ -1,0 +1,12 @@
+from minimd._about import __author__
+from minimd._about import __author_email__
+from minimd._about import __license__
+from minimd._about import __url__
+from minimd._about import __version__
+from minimd.tokenizer import tokenize_files
+
+
+__all__ = [
+    '__author__', '__author_email__', '__license__', '__url__', '__version__',
+    'tokenize_files',
+]
