+++ minimd/_about.py
@@ -1,5 +1,5 @@
 __author__ = 'entwanne'
 __author_email__ = 'antoine.rozo@gmail.com'
 __license__ = 'MIT'
-__url__ = 'https://github.com/entwanne/lucina/tree/minimd'
+__url__ = 'https://github.com/entwanne/minimd'
 __version__ = '0.0.1'
