+++ setup.cfg
@@ -4,7 +4,7 @@ testpaths = tests
 
 [isort]
 force_single_line=True
-known_lucina=lucina
+known_minimd=minimd
 known_tests=tests
 sections=FUTURE,STDLIB,THIRDPARTY,CORE,FIRSTPARTY,TESTS,LOCALFOLDER
 default_section=THIRDPARTY
